requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
lxml==4.9.3
opencv-python>=4.5.0
pyzbar>=0.1.8
numpy>=1.19.0
Pillow>=8.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
openai>=1.0.0
python-dotenv>=1.0.0
